refactor: folder controller
