+++ src/routes/api/Transaction.ts
@@ -1,6 +1,6 @@
 import * as express from 'express';
 
-import controller from '../../controllers/Transaction';
+import controller from '../../controllers/api/Transaction';
 import { asyncMiddleware } from '../../plugins/utility';
 import Validator from '../../validator/util';
